Break the bfcache (aka page cache) in safari
Firefox will not put a page in the bfcache if it has "cache-control: no-store".
Safari will not put a page in the bfcache if it has "cache-control: no-store" AND is served as https.
Unfortunately we test our site on localhost with http.
Safari also ignores the unload handler trick so don't try that.